% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_gl_rule}
\alias{get_gl_rule}
\title{Get Gauss–Legendre Quadrature Nodes and Weights}
\usage{
get_gl_rule(n)
}
\arguments{
\item{n}{number of nodes.}
}
\description{
Get Gauss–Legendre Quadrature Nodes and Weights
}
\examples{
library(SimSurvNMarker)
get_gl_rule(4)
get_gl_rule(25)

}
