% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_marker}
\alias{sim_marker}
\title{Simulate a Number of Observed Marker for an Individual}
\usage{
sim_marker(B, U, sigma_chol, r_n_marker, r_obs_time, m_func, g_func, offset)
}
\arguments{
\item{B}{coefficient matrix for time-varying fixed effects. Use \code{NULL} if there is no effect.}

\item{U}{random effect matrix for time-varying random effects. Use \code{NULL} if there is no effect.}

\item{sigma_chol}{Cholesky decomposition of the noise's covariance matrix.}

\item{r_n_marker}{function to generate the number of observed markers.}

\item{r_obs_time}{function to generate the observations times given the
number of observed markers. It should take a single
integer.}

\item{m_func}{basis function for \code{U} like \code{\link{poly}}.}

\item{g_func}{basis function for \code{B} like \code{\link{poly}}.}

\item{offset}{vector with non-time-varying fixed effects.}
}
\description{
Simulate a Number of Observed Marker for an Individual
}
