% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{surv_func_joint}
\alias{surv_func_joint}
\title{Evaluate the Conditional Survival Function Given the Random Effect}
\usage{
surv_func_joint(
  ti,
  B,
  U,
  omega,
  delta,
  alpha,
  b_func,
  m_func,
  gl_dat = get_gl_rule(30L),
  g_func,
  offset
)
}
\arguments{
\item{ti}{time points.}

\item{B}{coefficient matrix for time-varying fixed effects. Use \code{NULL} if there is no effect.}

\item{U}{random effect matrix for time-varying random effects. Use \code{NULL} if there is no effect.}

\item{omega}{coefficient for baseline hazard.}

\item{delta}{offset on the log hazard scale. Use \code{NULL} if there is no effect.}

\item{alpha}{numeric vector with association parameters.}

\item{b_func}{basis function for baseline hazard like \code{\link{poly}}.}

\item{m_func}{basis function for \code{U} like \code{\link{poly}}.}

\item{gl_dat}{Gauss–Legendre quadrature data.}

\item{g_func}{basis function for \code{B} like \code{\link{poly}}.}

\item{offset}{vector with non-time-varying fixed effects.}
}
\description{
Evaluate the Conditional Survival Function Given the Random Effect
}
