% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{eval_surv_base_fun}
\alias{eval_surv_base_fun}
\title{Evalutes the Survival Function without a Marker}
\usage{
eval_surv_base_fun(
  ti,
  omega,
  b_func,
  gl_dat = SimSurvNMarker:::get_gl_rule(30L),
  delta = 0
)
}
\description{
Evalutes the Survival Function without a Marker
}
