% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_joint_data_set}
\alias{sim_joint_data_set}
\title{Simulate Individuals from a Joint Survival and Marker Model}
\usage{
sim_joint_data_set(
  n_obs,
  B,
  Psi,
  omega,
  delta,
  alpha,
  sigma,
  gamma,
  b_func,
  m_func,
  g_func,
  gl_dat = get_gl_rule(30L),
  r_z,
  r_left_trunc,
  r_right_cens,
  r_n_marker,
  r_x,
  r_obs_time,
  y_max,
  tol = .Machine$double.eps^(1/4)
)
}
\arguments{
\item{n_obs}{integer with the number of individuals to draw.}

\item{B}{coefficient matrix for time-varying fixed effects.
Use \code{NULL} if there is no effect.}

\item{Psi}{the random effects' covariance matrix.}

\item{omega}{numeric vector with coefficient for baseline hazard.}

\item{delta}{offset on the log hazard scale. Use \code{NULL} if
there is no effect.}

\item{alpha}{numeric vector with association parameters.}

\item{sigma}{the noise's covariance matrix.}

\item{gamma}{coefficient matrix for the non-time-varying fixed effects.
Use \code{NULL} if there is no effect.}

\item{b_func}{basis function for baseline hazard like \code{\link{poly}}.}

\item{m_func}{basis function for \code{U} like \code{\link{poly}}.}

\item{g_func}{basis function for \code{B} like \code{\link{poly}}.}

\item{gl_dat}{Gauss–Legendre quadrature data.}

\item{r_z}{generator for the covariates in the log hazard. Takes an
integer for the individual's id.}

\item{r_left_trunc}{generator for the left-truncation time.
Takes an integer for the individual's id.}

\item{r_right_cens}{generator for the right-censoring time. Takes an
integer for the individual's id.}

\item{r_n_marker}{function to generate the number of observed markers.
Takes an integer for the individual's id.}

\item{r_x}{generator for the covariates in for the markers. Takes an
integer for the individual's id.}

\item{r_obs_time}{function to generate the observations times given the
number of observed markers. Takes an
integer for the number of markers and an integer
for the individual's id.}

\item{y_max}{maximum survival time before administrative censoring.}

\item{tol}{convergence tolerance passed to \code{\link{uniroot}}.}
}
\description{
Simulates individuals from the following model

\deqn{\vec U_i \sim N^{(K)}(\vec 0, \Psi)}
\deqn{\vec Y_{ij} \mid \vec U_i = \vec u_i \sim N^{(r)}(\vec \mu_i(s_{ij}, \vec u_i), \Sigma)}
\deqn{h(t \mid \vec u) = \exp(\vec\omega^\top\vec b(t) + \delta +
  \vec 1^\top(diag(\vec \alpha) \otimes \vec x_i^\top)vec(\Gamma) +
  \vec 1^\top(diag(\vec \alpha) \otimes \vec g(t)^\top)vec(B) +
  \vec 1^\top(diag(\vec \alpha) \otimes \vec m(t)^\top)\vec u).}

with

\deqn{\vec\mu_i(s, \vec u) = (I \otimes \vec x_i^\top)vec(\Gamma) + \left(I \otimes \vec g(s)^\top\right)vec(B) + \left(I \otimes \vec m(s)^\top\right) \vec u}

where \eqn{h(t \mid \vec u)} is the conditional hazard function.
}
\seealso{
See the examples on Github at
\url{https://github.com/boennecd/SimSurvNMarker} and
\url{https://github.com/boennecd/SimSurvNMarker/tree/master/inst/test-data}.

\code{\link{sim_marker}}, \code{\link{surv_func_joint}}
}
