% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_joint_data_set}
\alias{sim_joint_data_set}
\title{Simulate Individuals from a Joint Survival and Marker Model}
\usage{
sim_joint_data_set(
  n_obs,
  B,
  Psi,
  omega,
  delta,
  alpha,
  sigma,
  gamma,
  b_func,
  m_func,
  g_func,
  gl_dat = get_gl_rule(30L),
  r_z,
  r_left_trunc,
  r_right_cens,
  r_n_marker,
  r_x,
  r_obs_time,
  y_max
)
}
\arguments{
\item{n_obs}{integer with the number of individuals to draw.}

\item{B}{coefficient matrix for time-varying fixed effects. Use \code{NULL} if there is no effect.}

\item{Psi}{the random effect covariance matrix.}

\item{omega}{coefficient for baseline hazard.}

\item{delta}{offset on the log hazard scale. Use \code{NULL} if there is no effect.}

\item{alpha}{numeric vector with association parameters.}

\item{sigma}{the noise's covariance matrix.}

\item{gamma}{coefficient matrix for the non-time-varying fixed effects. Use \code{NULL} if there is no effect.}

\item{b_func}{basis function for baseline hazard like \code{\link{poly}}.}

\item{m_func}{basis function for \code{U} like \code{\link{poly}}.}

\item{g_func}{basis function for \code{B} like \code{\link{poly}}.}

\item{gl_dat}{Gauss–Legendre quadrature data.}

\item{r_z}{generator for the covariates in the log hazard. Takes an integer id.}

\item{r_left_trunc}{generator for the left-truncation time.
Takes an integer for the individual's id.}

\item{r_right_cens}{generator for the right-censoring time. Takes an
integer for the individual's id.}

\item{r_n_marker}{function to generate the number of observed markers.
Takes an integer for the individual's id.}

\item{r_x}{generator for the covariates in for the markers. Takes an
integer for the individual's id.}

\item{r_obs_time}{function to generate the observations times given the
number of observed markers. Takes an
integer for the number of markers and an integer
for the individual's id.}

\item{y_max}{maximum survival time before administrative censoring.}
}
\description{
Simulate Individuals from a Joint Survival and Marker Model
}
