% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_marker}
\alias{sim_marker}
\title{Simulate a Number of Observed Marker for an Individual}
\usage{
sim_marker(
  B,
  U,
  sigma_chol,
  r_n_marker,
  r_obs_time,
  m_func,
  g_func,
  offset,
  id = 1L
)
}
\arguments{
\item{B}{coefficient matrix for time-varying fixed effects.
Use \code{NULL} if there is no effect.}

\item{U}{random effects matrix for time-varying random effects.
Use \code{NULL} if there is no effects.}

\item{sigma_chol}{Cholesky decomposition of the noise's covariance matrix.}

\item{r_n_marker}{function to generate the number of observed markers.
Takes an integer for the individual's id.}

\item{r_obs_time}{function to generate the observations times given the
number of observed markers. Takes an
integer for the number of markers and an integer
for the individual's id.}

\item{m_func}{basis function for \code{U} like \code{\link{poly}}.}

\item{g_func}{basis function for \code{B} like \code{\link{poly}}.}

\item{offset}{numeric vector with non-time-varying fixed effects.}

\item{id}{integer with id passed to \code{r_n_marker} and
\code{r_obs_time}.}
}
\description{
Simulates from

\deqn{\vec U_i \sim N^{(K)}(\vec 0, \Psi)}
\deqn{\vec Y_{ij} \mid \vec U_i = \vec u_i \sim N^{(r)}(\vec \mu(s_{ij}, \vec u_i), \Sigma)}

with

\deqn{\vec\mu(s, \vec u) = \vec o + \left(I \otimes \vec g(s)^\top\right)vec(B) + \left(I \otimes \vec m(s)^\top\right) \vec u.}

The number of observations and the observations times, \eqn{s_{ij}}s, are
determined from the passed generating functions.
}
\seealso{
\code{\link{draw_U}}, \code{\link{eval_marker}}
}
