% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{surv_func_joint}
\alias{surv_func_joint}
\title{Evaluate the Conditional Survival Function Given the Random Effects}
\usage{
surv_func_joint(
  ti,
  B,
  U,
  omega,
  delta,
  alpha,
  b_func,
  m_func,
  gl_dat = get_gl_rule(30L),
  g_func,
  offset
)
}
\arguments{
\item{ti}{numeric vector with time points.}

\item{B}{coefficient matrix for time-varying fixed effects.
Use \code{NULL} if there is no effect.}

\item{U}{random effects matrix for time-varying random effects.
Use \code{NULL} if there is no effects.}

\item{omega}{numeric vector with coefficient for baseline hazard.}

\item{delta}{offset on the log hazard scale. Use \code{NULL} if
there is no effect.}

\item{alpha}{numeric vector with association parameters.}

\item{b_func}{basis function for baseline hazard like \code{\link{poly}}.}

\item{m_func}{basis function for \code{U} like \code{\link{poly}}.}

\item{gl_dat}{Gauss–Legendre quadrature data.}

\item{g_func}{basis function for \code{B} like \code{\link{poly}}.}

\item{offset}{numeric vector with non-time-varying fixed effects.}
}
\description{
Evaluate the conditional survival function given the random effects,
\eqn{\vec U}. The conditional hazard function is

\deqn{h(t \mid \vec u) = \exp(\vec\omega^\top\vec b(t) + \delta +
  \vec\alpha^\top\vec o +
  \vec 1^\top(diag(\vec \alpha) \otimes \vec g(t)^\top)vec(B) +
  \vec 1^\top(diag(\vec \alpha) \otimes \vec m(t)^\top)\vec u).}
}
\seealso{
\code{\link{sim_marker}}, \code{\link{draw_U}},
\code{\link{eval_surv_base_fun}}
}
